requirements.txt
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
av==14.2.0
blessed==1.20.0
blobfile==3.0.0
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///croot/certifi_1738623731865/work/certifi
charset-normalizer==3.4.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cramjam==2.9.1
cycler==0.12.1
datasets==3.5.0
decorator==5.2.1
decord==0.6.0
deepspeed==0.16.4
diffusers==0.32.2
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dotenv==0.9.9
easyocr==1.7.2
einops==0.8.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
EzColorLog==1.0.3
fairscale==0.4.13
fastparquet==2024.11.0
filelock==3.18.0
FLAML==2.3.4
flash_attn==2.7.4.post1
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.6.0
fsspec==2024.12.0
ftfy==6.3.1
gmpy2 @ file:///croot/gmpy2_1738085463648/work
gpustat==1.1.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
iopath==0.1.10
ipython==8.34.0
jedi==0.19.2
Jinja2 @ file:///croot/jinja2_1741710844255/work
jiter==0.9.0
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazy_loader==0.4
lightgbm==4.6.0
llvmlite==0.44.0
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
networkx @ file:///croot/networkx_1737039604450/work
ninja==1.11.1.4
numba==0.61.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.4.2
omegaconf==2.3.0
onnxruntime==1.21.1
open_clip_torch==2.31.0
openai==1.68.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.15.2
pexpect==4.9.0
pillow @ file:///croot/pillow_1738010226202/work
platformdirs==4.3.7
pooch==1.8.2
portalocker==3.1.1
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyclipper==1.3.0.post6
pycryptodomex==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyMatting==1.1.13
pyparsing==3.2.3
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
qwen-vl-utils==0.0.10
referencing==0.36.2
regex==2024.11.6
rembg==2.0.65
requests @ file:///croot/requests_1730999120400/work
rich==13.9.4
rpds-py==0.24.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sty==1.0.6
style==1.1.0
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.9.0
timeout-decorator==0.5.0
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.4.1
torchvision==0.21.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.2.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3 @ file:///croot/urllib3_1737133630106/work
validators==0.34.0
wcwidth==0.2.13
xformers==0.0.29.post3
xgboost==2.1.4
XlsxWriter==3.2.2
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
